Move to common method to convert the form data to DomainForwardingSaveData
