+++ client/dashboard/domains/domain-forwarding/add.tsx
@@ -8,7 +8,7 @@ import { domainRoute, domainForwardingsRoute } from '../../app/router';
 import { PageHeader } from '../../components/page-header';
 import PageLayout from '../../components/page-layout';
 import DomainForwardingForm from './form';
-import { parseTargetUrl } from './utils';
+import { formDataToSubmitData } from './utils';
 import type { FormData } from './form';
 import type { DomainForwardingSaveData } from '../../data/domain-forwarding';
 
@@ -19,17 +19,7 @@ export default function AddDomainForwarding() {
 	const { createSuccessNotice, createErrorNotice } = useDispatch( noticesStore );
 
 	const handleSubmit = ( formData: FormData ) => {
-		const { target_host, target_path, is_secure } = parseTargetUrl( formData.targetUrl );
-
-		const submitData: DomainForwardingSaveData = {
-			subdomain:
-				formData.sourceType === 'subdomain' ? formData.subdomain.trim() || undefined : undefined,
-			target_host,
-			target_path,
-			is_secure,
-			is_permanent: formData.redirectType === 'permanent',
-			forward_paths: formData.pathForwarding === 'yes',
-		};
+		const submitData: DomainForwardingSaveData = formDataToSubmitData( formData );
 
 		saveMutation.mutate( submitData, {
 			onSuccess: () => {
